btrfs-progs: extract btrfs_link_subvol from btrfs_mksubvol
The original btrfs_mksubvol is too specific to specify the directory that
the subvolume will link to. Furthermore, in this transaction, we don't only
need to create root_ref/dir-item, but also update the flags of root_item.
Extract a generic btrfs_link_subvol that allow the caller pass a trans
argument for later subvolume undelete. And rename the original
btrfs_mksubvol to link_subvol_for_convert since it is a less generic
btrfs_link_subvol.

No functional changes.

Signed-off-by: Lu Fengqi <lufq.fnst@cn.fujitsu.com>